requirements.txt
transformers>=4.35.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
ultralytics>=8.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
streamlit>=1.28.0
gradio>=4.0.0
